+++ .storybook/preview.js
@@ -1,5 +1,5 @@
 import { setCustomElementsManifest } from '@storybook/web-components';
-import '@webcomponents/scoped-custom-element-registry';
+import '@muonic/muon/js/scoped-custom-element-registry.min.js';
 import cem from '../dist/custom-elements.json';
 
 // import '../dist/muon.min.css';