+++ packages/muon/tests/components/inputter/__snapshots__/inputter.test.snap.js
@@ -1,10 +1,8 @@
 /* @web/test-runner snapshot v1 */
 export const snapshots = {};
+
 snapshots["Inputter standard default"] = 
-`<div
-  class="inputter"
-  style=""
->
+`<div class="inputter">
   <slot name="label">
   </slot>
   <div class="wrapper">
@@ -16,10 +14,7 @@ snapshots["Inputter standard default"] =
 /* end snapshot Inputter standard default */
 
 snapshots["Inputter helper text"] = 
-`<div
-  class="inputter"
-  style=""
->
+`<div class="inputter type-text">
   <slot name="label">
   </slot>
   <div class="helper">
@@ -34,10 +29,7 @@ snapshots["Inputter helper text"] =
 /* end snapshot Inputter helper text */
 
 snapshots["Inputter helper detail"] = 
-`<div
-  class="inputter"
-  style=""
->
+`<div class="inputter type-text">
   <slot name="label">
   </slot>
   <inputter-detail>
@@ -56,10 +48,7 @@ snapshots["Inputter helper detail"] =
 /* end snapshot Inputter helper detail */
 
 snapshots["Inputter helper detail open"] = 
-`<div
-  class="inputter"
-  style=""
->
+`<div class="inputter type-text">
   <slot name="label">
   </slot>
   <inputter-detail open="">
@@ -77,11 +66,20 @@ snapshots["Inputter helper detail open"] =
 `;
 /* end snapshot Inputter helper detail open */
 
+snapshots["Inputter text disabled"] = 
+`<div class="has-disabled inputter type-text">
+  <slot name="label">
+  </slot>
+  <div class="wrapper">
+    <slot>
+    </slot>
+  </div>
+</div>
+`;
+/* end snapshot Inputter text disabled */
+
 snapshots["Inputter text mask text"] = 
-`<div
-  class="has-mask inputter"
-  style="--maxlength:4;"
->
+`<div class="has-mask inputter type-text">
   <slot name="label">
   </slot>
   <div class="wrapper">
@@ -98,11 +96,8 @@ snapshots["Inputter text mask text"] =
 `;
 /* end snapshot Inputter text mask text */
 
-snapshots["Inputter text validation"] = 
-`<div
-  class="inputter"
-  style=""
->
+snapshots["Inputter text validation on input"] = 
+`<div class="inputter type-text">
   <slot name="label">
   </slot>
   <div class="wrapper">
@@ -111,29 +106,10 @@ snapshots["Inputter text validation"] =
   </div>
 </div>
 `;
-/* end snapshot Inputter text validation */
+/* end snapshot Inputter text validation on input */
 
-snapshots["Inputter radio standard radio"] = 
-`<div
-  class="inputter radio"
-  style=""
->
-  <span class="input-heading">
-    What is your heating source?
-  </span>
-  <div class="wrapper">
-    <slot>
-    </slot>
-  </div>
-</div>
-`;
-/* end snapshot Inputter radio standard radio */
-
-snapshots["Inputter standard default"] = 
-`<div
-  class="inputter"
-  style=""
->
+snapshots["Inputter text validation on change"] = 
+`<div class="inputter type-text">
   <slot name="label">
   </slot>
   <div class="wrapper">
@@ -142,141 +118,10 @@ snapshots["Inputter standard default"] =
   </div>
 </div>
 `;
-/* end snapshot Inputter standard default */
-
-snapshots["Inputter helper text"] = 
-`<div
-  class="inputter"
-  style=""
->
-  <slot name="label">
-  </slot>
-  <div class="helper">
-    What is this?
-  </div>
-  <div class="wrapper">
-    <slot>
-    </slot>
-  </div>
-</div>
-`;
-/* end snapshot Inputter helper text */
-
-snapshots["Inputter helper detail"] = 
-`<div
-  class="inputter"
-  style=""
->
-  <slot name="label">
-  </slot>
-  <inputter-detail>
-    <div slot="heading">
-      What is this?
-    </div>
-    <slot name="tip-details">
-    </slot>
-  </inputter-detail>
-  <div class="wrapper">
-    <slot>
-    </slot>
-  </div>
-</div>
-`;
-/* end snapshot Inputter helper detail */
-
-snapshots["Inputter helper detail open"] = 
-`<div
-  class="inputter"
-  style=""
->
-  <slot name="label">
-  </slot>
-  <inputter-detail open="">
-    <div slot="heading">
-      What is this?
-    </div>
-    <slot name="tip-details">
-    </slot>
-  </inputter-detail>
-  <div class="wrapper">
-    <slot>
-    </slot>
-  </div>
-</div>
-`;
-/* end snapshot Inputter helper detail open */
-
-snapshots["Inputter select standard select"] = 
-`<div
-  class="inputter select"
-  style=""
->
-  <slot name="label">
-  </slot>
-  <div class="wrapper">
-    <slot>
-    </slot>
-    <inputter-icon name="chevron-circle-down">
-    </inputter-icon>
-  </div>
-</div>
-`;
-/* end snapshot Inputter select standard select */
-
-snapshots["Inputter search standard search"] = 
-`<div
-  class="inputter search"
-  style=""
->
-  <slot name="label">
-  </slot>
-  <div class="wrapper">
-    <slot>
-    </slot>
-    <inputter-icon name="search">
-    </inputter-icon>
-  </div>
-</div>
-`;
-/* end snapshot Inputter search standard search */
-
-snapshots["Inputter date standard date"] = 
-`<div
-  class="date inputter"
-  style=""
->
-  <slot name="label">
-  </slot>
-  <div class="wrapper">
-    <slot>
-    </slot>
-    <inputter-icon name="calendar">
-    </inputter-icon>
-  </div>
-</div>
-`;
-/* end snapshot Inputter date standard date */
-
-snapshots["Inputter text disabled"] = 
-`<div
-  class="has-disabled inputter"
-  style=""
->
-  <slot name="label">
-  </slot>
-  <div class="wrapper">
-    <slot>
-    </slot>
-  </div>
-</div>
-`;
-/* end snapshot Inputter text disabled */
+/* end snapshot Inputter text validation on change */
 
 snapshots["Inputter text mask & validation"] = 
-`<div
-  class="has-mask inputter"
-  style="--maxlength:4;"
->
+`<div class="has-mask inputter type-text">
   <slot name="label">
   </slot>
   <div class="wrapper">
@@ -293,11 +138,21 @@ snapshots["Inputter text mask & validation"] =
 `;
 /* end snapshot Inputter text mask & validation */
 
+snapshots["Inputter radio standard radio"] = 
+`<div class="inputter radio">
+  <span class="input-heading">
+    What is your heating source?
+  </span>
+  <div class="wrapper">
+    <slot>
+    </slot>
+  </div>
+</div>
+`;
+/* end snapshot Inputter radio standard radio */
+
 snapshots["Inputter radio radio mask"] = 
-`<div
-  class="has-mask inputter radio"
-  style="--maxlength:3;"
->
+`<div class="has-mask inputter radio">
   <span class="input-heading">
     What is your heating source?
   </span>
@@ -310,10 +165,7 @@ snapshots["Inputter radio radio mask"] =
 /* end snapshot Inputter radio radio mask */
 
 snapshots["Inputter radio radio mask validation"] = 
-`<div
-  class="has-mask inputter radio"
-  style="--maxlength:3;"
->
+`<div class="has-mask inputter radio">
   <span class="input-heading">
     What is your heating source?
   </span>
@@ -325,33 +177,63 @@ snapshots["Inputter radio radio mask validation"] =
 `;
 /* end snapshot Inputter radio radio mask validation */
 
-snapshots["Inputter text validation on input"] = 
-`<div
-  class="inputter"
-  style=""
->
+snapshots["Inputter select standard select"] = 
+`<div class="inputter select">
   <slot name="label">
   </slot>
   <div class="wrapper">
     <slot>
     </slot>
+    <inputter-icon name="chevron-circle-down">
+    </inputter-icon>
   </div>
 </div>
 `;
-/* end snapshot Inputter text validation on input */
+/* end snapshot Inputter select standard select */
 
-snapshots["Inputter text validation on change"] = 
-`<div
-  class="inputter"
-  style=""
->
+snapshots["Inputter search standard search"] = 
+`<div class="inputter search type-search">
   <slot name="label">
   </slot>
   <div class="wrapper">
     <slot>
     </slot>
+    <inputter-icon name="search">
+    </inputter-icon>
   </div>
 </div>
 `;
-/* end snapshot Inputter text validation on change */
+/* end snapshot Inputter search standard search */
+
+snapshots["Inputter date standard date"] = 
+`<div class="date inputter type-date">
+  <slot name="label">
+  </slot>
+  <div class="wrapper">
+    <slot>
+    </slot>
+    <inputter-icon name="calendar">
+    </inputter-icon>
+  </div>
+</div>
+`;
+/* end snapshot Inputter date standard date */
+
+snapshots["Inputter text mask & separator text"] = 
+`<div class="has-mask inputter type-text">
+  <slot name="label">
+  </slot>
+  <div class="wrapper">
+    <slot>
+    </slot>
+    <div
+      aria-hidden="true"
+      class="input-mask"
+    >
+      00-00
+    </div>
+  </div>
+</div>
+`;
+/* end snapshot Inputter text mask & separator text */