+++ packages/muon/tests/scripts/utils/cta-component.js
@@ -1,0 +1,17 @@
+import { MuonElement } from '@muonic/muon';
+
+/**
+ * A Nucleus call-to-action allows users to take action once they are ready for it.
+ *
+ * @element cta
+ * @prefix mnx
+ *
+ */
+
+export class TestCta extends MuonElement {
+  static get properties() {
+    return {
+      enabled: { type: Boolean }
+    };
+  }
+}